+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # assert.js version history
 
+## assert.js v1.0.3
+
+1. Documentation, pdf and code fixes.
+2. Add inner links in the __readme.html__.
+3. Add 2 new properties of the TestResult object (`block`, `name`) in the testrunner functions.
+4. Add these functions: `inRange();`, `notInRange();`
+
 ## assert.js v1.0.2
 
 1. Documentation, pdf and code fixes.